Moving logger to dapr-ext-workflow
